+++ packages/huggingface/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @ai-sdk/huggingface
 
+## 1.0.0-beta.30
+
+### Patch Changes
+
+- 1bbce64: extract reasoning content
+
+## 1.0.0-beta.29
+
+### Patch Changes
+
+- Updated dependencies [b681d7d]
+  - @ai-sdk/provider@3.0.0-beta.16
+  - @ai-sdk/openai-compatible@2.0.0-beta.34
+  - @ai-sdk/provider-utils@4.0.0-beta.33
+
+## 1.0.0-beta.28
+
+### Patch Changes
+
+- Updated dependencies [32d8dbb]
+  - @ai-sdk/provider-utils@4.0.0-beta.32
+  - @ai-sdk/openai-compatible@2.0.0-beta.33
+
 ## 1.0.0-beta.27
 
 ### Patch Changes