+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,65 @@
 
 ### Patch Changes
 
+- ai@6.0.0-beta.103
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@6.0.0-beta.102
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@6.0.0-beta.101
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [8370068]
+  - ai@6.0.0-beta.100
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [384142c]
+  - ai@6.0.0-beta.99
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [b681d7d]
+  - ai@6.0.0-beta.98
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@6.0.0-beta.97
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [a322efa]
+  - ai@6.0.0-beta.96
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [eb8d1cb]
+  - ai@6.0.0-beta.95
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
 - Updated dependencies [ab1087b]
   - ai@6.0.0-beta.94